requirements.txt
appdirs==1.4.3
asgiref==3.2.10
CacheControl==0.12.6
cachetools==4.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.7
django-appconf==1.0.4
django-bootstrap3==14.0.0
django-braces==1.14.0
django-cleanup==5.0.0
django-debug-toolbar==2.2
django-googledrive-storage==1.5.1
django-heroku==0.3.1
djangorestframework==3.11.0
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.18.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
misaka==2.1.1
msgpack==0.6.2
oauth2client==4.1.3
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.6
python-dateutil==2.8.1
python-memcached==1.59
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
rsa==4.6
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.1.0
